Date Time Provider Refactor
